+++ EssentialApp/EssentialApp/SceneDelegate.swift
@@ -11,8 +11,11 @@ import EssentialFeed
 import EssentialFeediOS
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
-  
   var window: UIWindow?
+  
+  let localStoreURL = NSPersistentContainer
+    .defaultDirectoryURL()
+    .appendingPathComponent("feed-store.sqlite")
   
   func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
     guard let _ = (scene as? UIWindowScene) else { return }
@@ -22,16 +25,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     let remoteClient = makeRemoteClient()
     let remotefeedLoader = RemoteFeedLoader(url: remoteURL, client: remoteClient)
     let remoteImageLoader = RemoteFeedImageDataLoader(client: remoteClient)
-    
-    let localStoreURL = NSPersistentContainer
-      .defaultDirectoryURL()
-      .appendingPathComponent("feed-store.sqlite")
-    
-    #if DEBUG
-    if CommandLine.arguments.contains(where: { $0 == "-reset"}) {
-      try? FileManager.default.removeItem(at: localStoreURL)
-    }
-    #endif
     
     let localStore = try! CoreDataFeedStore(storeURL: localStoreURL)
     let localFeedLoader = LocalFeedLoader(store: localStore, currentDate: Date.init)
@@ -50,26 +43,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
           cache: localImageDataLoader)))
   }
   
-  private func makeRemoteClient() -> HTTPClient {
-    #if DEBUG
-    if UserDefaults.standard.string(forKey: "connectivity") == "offline" {
-      return AlwaysFalingHTTPClient()
-    }
-    #endif
-    
+  func makeRemoteClient() -> HTTPClient {
     return URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
   }
 }
-
-#if DEBUG
-private class AlwaysFalingHTTPClient: HTTPClient {
-  private class Task: HTTPClientTask {
-    func cancel() {}
-  }
-  
-  func get(from url: URL, completion: @escaping (HTTPClient.Result) -> Void) -> HTTPClientTask {
-    completion(.failure(NSError(domain: "offline", code: 0)))
-    return Task()
-  }
-}
-#endif
